+++ packages/preview-middleware-client/src/adp/api-handler.ts
@@ -4,6 +4,7 @@ export const enum ApiEndpoints {
     CHANGES = '/preview/api/changes',
     FRAGMENT = '/adp/api/fragment',
     CONTROLLER = '/adp/api/controller',
+    CODE_EXT = '/adp/api/code_ext',
     MANIFEST_APP_DESCRIPTOR = '/manifest.appdescr_variant'
 }
 
@@ -21,6 +22,12 @@ type Controllers = { controllerName: string }[];
 export interface FragmentsResponse {
     fragments: Fragments;
     message: string;
+}
+
+export interface CodeExtResponse {
+    controllerExists: boolean;
+    controllerPath: string;
+    controllerPathFromRoot: string;
 }
 
 export interface ControllersResponse {
@@ -60,7 +67,9 @@ export async function request<T>(endpoint: ApiEndpoints, method: RequestMethod, 
         const response: Response = await fetch(endpoint, config);
 
         if (!response.ok) {
-            throw new Error(`Request failed, status: ${response.status}.`);
+            const errorData = await response.json();
+            const message = errorData?.message ?? '';
+            throw new Error(`Request failed, status: ${response.status}. ${message}`.trim());
         }
 
         switch (method) {
@@ -125,6 +134,16 @@ export async function readControllers<T>(): Promise<T> {
  */
 export async function writeController<T>(data: T): Promise<T> {
     return request<T>(ApiEndpoints.CONTROLLER, RequestMethod.POST, data);
+}
+
+/**
+ * Checks for existing controller in the project's workspace
+ *
+ * @param controllerName Name of the controller
+ * @returns {CodeExtResponse} Returns path to existing controller if found
+ */
+export async function getExistingController(controllerName: string): Promise<CodeExtResponse> {
+    return request<CodeExtResponse>(`${ApiEndpoints.CODE_EXT}/${controllerName}` as ApiEndpoints, RequestMethod.GET);
 }
 
 /**